Moved db,js file
